# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * VendorFlow API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MessageResponse {
  message: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  admin: "admin",
  manager: "manager",
  approver: "approver",
  user: "user",
} as const;

export interface User {
  id: number;
  email: string;
  name: string;
  role: UserRole;
  tenantId: number;
  createdAt: string;
}

export interface AuthResponse {
  user: User;
}

export interface StatusUpdate {
  status: string;
  /** @nullable */
  rejectionReason?: string | null;
}

export type VendorStatus = (typeof VendorStatus)[keyof typeof VendorStatus];

export const VendorStatus = {
  active: "active",
  inactive: "inactive",
  suspended: "suspended",
} as const;

export interface Vendor {
  id: number;
  name: string;
  email: string;
  phone: string;
  address: string;
  category: string;
  rating: number;
  totalSpent: number;
  status: VendorStatus;
  createdAt: string;
  updatedAt: string;
}

export type VendorInputStatus =
  (typeof VendorInputStatus)[keyof typeof VendorInputStatus];

export const VendorInputStatus = {
  active: "active",
  inactive: "inactive",
  suspended: "suspended",
} as const;

export interface VendorInput {
  name: string;
  email: string;
  phone: string;
  address: string;
  category: string;
  rating: number;
  status: VendorInputStatus;
}

export type VendorUpdateStatus =
  (typeof VendorUpdateStatus)[keyof typeof VendorUpdateStatus];

export const VendorUpdateStatus = {
  active: "active",
  inactive: "inactive",
  suspended: "suspended",
} as const;

export interface VendorUpdate {
  name?: string;
  email?: string;
  phone?: string;
  address?: string;
  category?: string;
  rating?: number;
  status?: VendorUpdateStatus;
}

export interface PurchaseOrderItem {
  id: number;
  description: string;
  quantity: number;
  unitPrice: number;
  total: number;
  category: string;
}

export interface PurchaseOrderItemInput {
  description: string;
  quantity: number;
  unitPrice: number;
  category: string;
}

export type PurchaseOrderStatus =
  (typeof PurchaseOrderStatus)[keyof typeof PurchaseOrderStatus];

export const PurchaseOrderStatus = {
  draft: "draft",
  submitted: "submitted",
  approved: "approved",
  rejected: "rejected",
  received: "received",
  cancelled: "cancelled",
} as const;

export interface PurchaseOrder {
  id: number;
  poNumber: string;
  vendorId: number;
  vendorName: string;
  items: PurchaseOrderItem[];
  totalAmount: number;
  status: PurchaseOrderStatus;
  /** @nullable */
  notes?: string | null;
  dueDate: string;
  createdBy: number;
  createdAt: string;
  updatedAt: string;
}

export interface PurchaseOrderInput {
  vendorId: number;
  items: PurchaseOrderItemInput[];
  notes?: string;
  dueDate: string;
}

export interface PurchaseOrderUpdate {
  vendorId?: number;
  items?: PurchaseOrderItemInput[];
  notes?: string;
  dueDate?: string;
}

export interface InventoryItem {
  id: number;
  sku: string;
  name: string;
  description: string;
  quantity: number;
  reorderLevel: number;
  warehouseLocation: string;
  unitCost: number;
  category: string;
  /** @nullable */
  vendorId?: number | null;
  /** @nullable */
  vendorName?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface InventoryItemInput {
  sku: string;
  name: string;
  description: string;
  quantity: number;
  reorderLevel: number;
  warehouseLocation: string;
  unitCost: number;
  category: string;
  /** @nullable */
  vendorId?: number | null;
}

export interface InventoryItemUpdate {
  sku?: string;
  name?: string;
  description?: string;
  quantity?: number;
  reorderLevel?: number;
  warehouseLocation?: string;
  unitCost?: number;
  category?: string;
  /** @nullable */
  vendorId?: number | null;
}

export type AuditLogChanges = { [key: string]: unknown };

export interface AuditLog {
  id: number;
  userId: number;
  userName: string;
  action: string;
  entityType: string;
  entityId: string;
  changes: AuditLogChanges;
  /** @nullable */
  ipAddress?: string | null;
  timestamp: string;
}

export interface DashboardStats {
  totalVendors: number;
  activeOrders: number;
  pendingApprovals: number;
  totalSpending: number;
  inventoryValue: number;
  lowStockCount: number;
}

export interface VendorSpending {
  vendorId: number;
  vendorName: string;
  totalSpent: number;
}

export interface StatusCount {
  status: string;
  count: number;
}

export type NotificationType =
  (typeof NotificationType)[keyof typeof NotificationType];

export const NotificationType = {
  approval_request: "approval_request",
  order_status: "order_status",
  inventory_alert: "inventory_alert",
  system: "system",
} as const;

export interface Notification {
  id: number;
  type: NotificationType;
  title: string;
  message: string;
  read: boolean;
  /** @nullable */
  relatedEntityId?: string | null;
  createdAt: string;
}