# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: VendorFlow API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /inventory:
    get:
      operationId: listInventory
      tags: [inventory]
      summary: List inventory items
      responses:
        "200":
          description: List of inventory items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InventoryItem"
    post:
      operationId: createInventoryItem
      tags: [inventory]
      summary: Create an inventory item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InventoryItemInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryItem"
  /inventory/{id}:
    get:
      operationId: getInventoryItem
      tags: [inventory]
      summary: Get an inventory item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Inventory item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryItem"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateInventoryItem
      tags: [inventory]
      summary: Update an inventory item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InventoryItemUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryItem"
    delete:
      operationId: deleteInventoryItem
      tags: [inventory]
      summary: Delete an inventory item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
  /notifications/{id}/read:
    patch:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark notification as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
  /notifications/{id}:
    delete:
      operationId: deleteNotification
      tags: [notifications]
      summary: Delete a notification
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
    LoginInput:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
      required:
        - user
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [admin, manager, approver, user]
        tenantId:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - email
        - name
        - role
        - tenantId
        - createdAt
...
    InventoryItem:
      type: object
      properties:
        id:
          type: integer
        sku:
          type: string
        name:
          type: string
        description:
          type: string
        quantity:
          type: number
        reorderLevel:
          type: number
        warehouseLocation:
          type: string
        unitCost:
          type: number
        category:
          type: string
        vendorId:
          type: ["integer", "null"]
        vendorName:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - sku
        - name
        - description
        - quantity
        - reorderLevel
        - warehouseLocation
        - unitCost
        - category
        - createdAt
        - updatedAt
    InventoryItemInput:
      type: object
      properties:
        sku:
          type: string
        name:
          type: string
        description:
          type: string
        quantity:
          type: number
        reorderLevel:
          type: number
        warehouseLocation:
          type: string
        unitCost:
          type: number
        category:
          type: string
        vendorId:
          type: ["integer", "null"]
      required:
        - sku
        - name
        - description
        - quantity
        - reorderLevel
        - warehouseLocation
        - unitCost
        - category
    InventoryItemUpdate:
      type: object
      properties:
        sku:
          type: string
        name:
          type: string
        description:
          type: string
        quantity:
          type: number
        reorderLevel:
          type: number
        warehouseLocation:
          type: string
        unitCost:
          type: number
        category:
          type: string
        vendorId:
          type: ["integer", "null"]
...
    Notification:
      type: object
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [approval_request, order_status, inventory_alert, system]
        title:
          type: string
        message:
          type: string
        read:
          type: boolean
        relatedEntityId:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - type
        - title
        - message
        - read
        - createdAt